变化： 1.在上一版本之上，解决了calendar与string的标准格式转换问题 2.将批量删除方法改为使用字符串拼接 3.在所有service层加入传入参数检查，返回为void函数检查未通过则不执行，有返回的函数未通过会返回null或者false 保证进入Dao层数据的干净 4.修改取操作者数据时显示的基本类型，以member类型为基准。展示的个人部门为显示第一个参加的部门 5.修复bug

+++ src/com/istc/Service/EntityService/MemberService.java
@@ -18,22 +18,35 @@ public class MemberService {
     private transient MemberDAO memberDAO;
 
     public Member get(String id){
+        if (id!=null)
         return memberDAO.get(id);
+        else
+            return null;
     }
     public void  update(Member member){
+        if (member!=null&&member.getID()!=null)
         memberDAO.edit(member);
     }
 
     public Member get(Member member){
-        if(member == null)return null;
+        if (member!=null&&member.getID()!=null)
         return memberDAO.get(member);
+        else return null;
+
     }
     public void remove(String id){
+        if (id!=null)
         memberDAO.delete(id);
+    }
+    public void remove(String[] ids){
+        if (ids!=null&&ids[0]!=null)
+            memberDAO.deleteSet(ids);
     }
 
     public boolean exist(Member member){
+        if (member!=null&&member.getID()!=null)
         return member == null ? false : get(member) != null;
+        else return false;
     }
     public boolean exist(String id){
         return id == null || "".equals(id) ? false : get(id) != null;
